requirements.txt
Click==7.0
dvc==0.40.1
gensim==3.7.2
gluonnlp==0.6.0
jieba-fast==0.53
mxnet-mkl==1.4.0
numpy==1.14.6
pandas==0.24.2
sanic==19.3.1
scikit-learn==0.20.3
scipy==1.2.1
sklearn==0.0
tqdm==4.31.1
ujson==1.35
oss2==2.6.1
